Feature/results

+++ src/main/kotlin/ru/erudyt/online/dto/request/CompetitionFilterRequest.kt
@@ -4,6 +4,6 @@ class CompetitionFilterRequest(
     val query: String?,
     val ageIds: List<Long>?,
     val subjectIds: List<Long>?,
-    val offset: Long?,
-    val limit: Long?,
+    val offset: Int?,
+    val limit: Int?,
 )
